requirements.txt
arrow==1.3.0
asgiref==3.7.2
astroid==3.0.3
blessed==1.20.0
certifi==2024.2.2
chardet==5.2.0
dj-database-url==2.1.0
Django==5.0.2
django-compat==1.0.15
django-extensions==3.2.3
django-picklefield==3.1
django-storages==1.14.2
dropbox==11.36.2
gunicorn==21.2.0
idna==3.6
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pylint==3.0.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typed-ast==1.5.5
urllib3==2.2.0
uWSGI==2.0.23
wcwidth==0.2.13
whitenoise==6.6.0
wrapt==1.16.0
django-widget-tweaks==1.5.0
